reorganized couple things
