--- webapp/hardhat.config 2.ts
@@ -1,8 +1,0 @@
-import { HardhatUserConfig } from "hardhat/config";
-import "@nomicfoundation/hardhat-toolbox";
-
-const config: HardhatUserConfig = {
-  solidity: "0.8.17",
-};
-
-export default config;
